Modify domain using fluents
Should work but no planner managed to solve it.

The domain is changed in a way that there are functions keeping track of
how many covers lie on each tile at any given time.

This means that putting on or off some cover is just increasing or
decreasing particular counters.
The information about how many covers are there on a given tile is then
used in color-applying actions, making the actions independent on the
covers themselves.
This makes it much easier to add new covers or change the existing ones.