Modify domain using types and actions parameters
Should work but no planner managed to solve it.

The domain is changed in a way that there is introduced type "color" and
it is used as the parameter for color-applying action.

Thanks to this modification, instead of many color-applying actions
there is just one now.
This removes redundant actions as well as redundant predicates,
generalizing the logic and letting particular problems to have custom
number of colors.